refactor: 테스트했던 코드 정리 및 재구성
