+++ api/src/main/java/com/unstage/api/auth/AuthController.java
@@ -1,11 +1,8 @@
-package com.unstage.api.auth.controller;
+package com.unstage.api.auth;
 
-import com.unstage.api.auth.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,27 +11,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/api/auth")
+@RequestMapping("/api/v1/auth")
 @RequiredArgsConstructor
 @Slf4j
 public class AuthController {
-
-    @GetMapping("/user")
-    public ResponseEntity<UserResponse> getCurrentUser(@AuthenticationPrincipal OidcUser oidcUser) {
-        if (oidcUser == null) {
-            return ResponseEntity.ok(UserResponse.builder()
-                    .authenticated(false)
-                    .build());
-        }
-
-        return ResponseEntity.ok(UserResponse.builder()
-                .authenticated(true)
-                .nickname(oidcUser.getAttribute("nickname"))
-                .email(oidcUser.getEmail())
-                .profileImageUrl(oidcUser.getAttribute("profile_image_url"))
-                .role(oidcUser.getAuthorities().iterator().next().getAuthority())
-                .build());
-    }
 
     @GetMapping("/login/success")
     public ResponseEntity<Map<String, String>> loginSuccess() {
