+++ api/src/main/java/com/unstage/api/auth/LoginController.java
@@ -1,4 +1,4 @@
-package com.unstage.api.auth.controller;
+package com.unstage.api.auth;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
