+++ api/src/main/java/com/unstage/api/auth/OidcUserInfo.java
@@ -1,4 +1,4 @@
-package com.unstage.api.auth.dto;
+package com.unstage.api.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
