+++ api/src/test/java/com/unstage/api/config/RestAssuredTest.java
@@ -1,8 +1,8 @@
 package com.unstage.api;
 
-
 import com.unstage.api.auth.config.TestSecurityConfig;
-import com.unstage.api.util.DatabaseCleanup;
+import com.unstage.core.testfixtures.DatabaseCleanup;
+import com.unstage.core.testfixtures.TestContainer;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,23 +11,13 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest(
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
-        , properties = {"spring.main.allow-bean-definition-overriding=true"}
-)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
 @EnableAutoConfiguration
-@ComponentScan(
-        basePackages = {"com.unstage.api", "com.unstage.core"},
-        excludeFilters = @ComponentScan.Filter(
-                type = FilterType.ASSIGNABLE_TYPE,
-                classes = com.unstage.api.config.SecurityConfig.class
-        )
-)
+@ComponentScan(basePackages = {"com.unstage.api", "com.unstage.core"})
 @Import(TestSecurityConfig.class)
 public abstract class RestAssuredTest extends TestContainer {
     @LocalServerPort
