+++ core/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
     implementation("org.springframework:spring-context")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-
     runtimeOnly("mysql:mysql-connector-java:${project.property('mysqlVersion')}")
 
     implementation("com.querydsl:querydsl-jpa:${project.property('querydslVersion')}:jakarta")
@@ -9,13 +8,6 @@ dependencies {
     annotationProcessor("jakarta.annotation:jakarta.annotation-api")
     annotationProcessor("jakarta.persistence:jakarta.persistence-api")
 
-    // 테스트 중
-    testImplementation("org.testcontainers:testcontainers:${project.property("testContainerVersion")}")
-    testImplementation("org.testcontainers:junit-jupiter:${project.property("testContainerVersion")}")
-    testImplementation("org.testcontainers:mysql:${project.property("testContainerMySqlVersion")}")
-    // 테스트 중
-
-    testFixturesApi("org.springframework.boot:spring-boot-starter-data-jpa")
     testFixturesImplementation("org.springframework.boot:spring-boot-starter-data-jpa")
     testFixturesImplementation("org.testcontainers:testcontainers:${project.property('testContainerVersion')}")
     testFixturesImplementation("org.testcontainers:junit-jupiter:${project.property('testContainerVersion')}")
