+++ core/src/testFixtures/java/com/unstage/core/testfixtures/TestContainer.java
@@ -1,4 +1,4 @@
-package com.unstage.api;
+package com.unstage.core.testfixtures;
 
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
